+++ ext/geos_c_impl/geometry.c
@@ -1131,25 +1131,57 @@ method_geometry_invalid_reason_location(VALUE self)
   return result;
 }
 
+typedef struct
+{
+  const GEOSGeometry* self_geom;
+  GEOSMakeValidParams* params;
+  GEOSGeometry* valid_geom;
+} make_valid_ctx;
+
 static VALUE
-method_geometry_make_valid(VALUE self)
+make_valid_perform(VALUE arg)
+{
+  make_valid_ctx* ctx = (make_valid_ctx*)arg;
+  ctx->valid_geom = GEOSMakeValidWithParams(ctx->self_geom, ctx->params);
+  return Qnil;
+}
+
+static VALUE
+make_valid_cleanup(VALUE arg)
+{
+  GEOSMakeValidParams_destroy(((make_valid_ctx*)arg)->params);
+  return Qnil;
+}
+
+static VALUE
+method_geometry_make_valid(VALUE self, VALUE method, VALUE keepCollapsed)
 {
   RGeo_GeometryData* self_data;
-  const GEOSGeometry* self_geom;
-  GEOSGeometry* valid_geom;
+  make_valid_ctx ctx;
+  int method_int;
+  int keep_collapsed_int;
+
   self_data = RGEO_GEOMETRY_DATA_PTR(self);
-  self_geom = self_data->geom;
-  if (!self_geom)
+  ctx.self_geom = self_data->geom;
+  if (!ctx.self_geom)
     return Qnil;
 
-  // According to GEOS implementation, MakeValid always returns.
-  valid_geom = GEOSMakeValid(self_geom);
-  if (!valid_geom) {
+  method_int = RB_NUM2INT(method);
+  keep_collapsed_int = RB_NUM2INT(keepCollapsed);
+
+  ctx.params = GEOSMakeValidParams_create();
+  GEOSMakeValidParams_setMethod(ctx.params, method_int);
+  GEOSMakeValidParams_setKeepCollapsed(ctx.params, keep_collapsed_int);
+  ctx.valid_geom = NULL;
+
+  rb_ensure(make_valid_perform, (VALUE)&ctx, make_valid_cleanup, (VALUE)&ctx);
+
+  if (!ctx.valid_geom) {
     rb_raise(rb_eRGeoInvalidGeometry,
              "%" PRIsVALUE,
              method_geometry_invalid_reason(self));
   }
-  return rgeo_wrap_geos_geometry(self_data->factory, valid_geom, Qnil);
+  return rgeo_wrap_geos_geometry(self_data->factory, ctx.valid_geom, Qnil);
 }
 
 static VALUE
@@ -1325,8 +1357,10 @@ rgeo_init_geos_geometry()
                    "point_on_surface",
                    method_geometry_point_on_surface,
                    0);
-  rb_define_method(
-    geos_geometry_methods, "make_valid", method_geometry_make_valid, 0);
+  rb_define_private_method(geos_geometry_methods,
+                           "geometry_make_valid",
+                           method_geometry_make_valid,
+                           2);
   rb_define_method(
     geos_geometry_methods, "polygonize", method_geometry_polygonize, 0);
 #ifdef RGEO_GEOS_SUPPORTS_DENSIFY